# perf: optimize `GetCombinations`