--- a/internal/util/relations.go
+++ b/internal/util/relations.go
@@ -9,47 +9,53 @@ type Rel struct {
 }
 
 func (relations *ForeignRelations) GetCombinations() []Rel {
+	var (
+		lConsumer      = len(relations.Consumer)
+		lConsumerGroup = len(relations.ConsumerGroup)
+		lRoutes        = len(relations.Route)
+		lServices      = len(relations.Service)
+		l              = lRoutes + lServices
+	)
+
 	var cartesianProduct []Rel
 
 	// gocritic I don't care that you think switch statements are the one true god of readability, the language offers
 	// multiple options for a reason. go away, gocritic.
-	if len(relations.Consumer) > 0 { //nolint:gocritic
-		consumers := relations.Consumer
-		if len(relations.Route)+len(relations.Service) > 0 {
-			for _, service := range relations.Service {
-				for _, consumer := range consumers {
+	if lConsumer > 0 { // nolint:gocritic
+		if l > 0 {
+			cartesianProduct = make([]Rel, 0, l*lConsumer)
+			for _, consumer := range relations.Consumer {
+				for _, service := range relations.Service {
 					cartesianProduct = append(cartesianProduct, Rel{
 						Service:  service,
 						Consumer: consumer,
 					})
 				}
-			}
-			for _, route := range relations.Route {
-				for _, consumer := range consumers {
+				for _, route := range relations.Route {
 					cartesianProduct = append(cartesianProduct, Rel{
 						Route:    route,
 						Consumer: consumer,
 					})
 				}
 			}
+
 		} else {
+			cartesianProduct = make([]Rel, 0, len(relations.Consumer))
 			for _, consumer := range relations.Consumer {
 				cartesianProduct = append(cartesianProduct, Rel{Consumer: consumer})
 			}
 		}
-	} else if len(relations.ConsumerGroup) > 0 {
-		groups := relations.ConsumerGroup
-		if len(relations.Route)+len(relations.Service) > 0 {
-			for _, service := range relations.Service {
-				for _, group := range groups {
+	} else if lConsumerGroup > 0 {
+		if l > 0 {
+			cartesianProduct = make([]Rel, 0, l*lConsumerGroup)
+			for _, group := range relations.ConsumerGroup {
+				for _, service := range relations.Service {
 					cartesianProduct = append(cartesianProduct, Rel{
 						Service:       service,
 						ConsumerGroup: group,
 					})
 				}
-			}
-			for _, route := range relations.Route {
-				for _, group := range groups {
+				for _, route := range relations.Route {
 					cartesianProduct = append(cartesianProduct, Rel{
 						Route:         route,
 						ConsumerGroup: group,
@@ -57,11 +63,13 @@ func (relations *ForeignRelations) GetCombinations() []Rel {
 				}
 			}
 		} else {
+			cartesianProduct = make([]Rel, 0, lConsumerGroup)
 			for _, group := range relations.ConsumerGroup {
 				cartesianProduct = append(cartesianProduct, Rel{ConsumerGroup: group})
 			}
 		}
-	} else {
+	} else if l > 0 {
+		cartesianProduct = make([]Rel, 0, l)
 		for _, service := range relations.Service {
 			cartesianProduct = append(cartesianProduct, Rel{Service: service})
 		}