**What this PR does / why we need it**:

Optimizes `GetCombinations` + adds a testcase for consumer groups

old e6de62fc1

```
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/util
BenchmarkGetCombinations/consumer_groups-10              3546044               346.5 ns/op           960 B/op          4 allocs/op
BenchmarkGetCombinations/consumer_groups-10              3691382               338.9 ns/op           960 B/op          4 allocs/op
BenchmarkGetCombinations/consumer_groups-10              3682710               366.5 ns/op           960 B/op          4 allocs/op
BenchmarkGetCombinations/consumer_groups-10              3747121               342.3 ns/op           960 B/op          4 allocs/op
BenchmarkGetCombinations/consumer_groups-10              3453127               331.8 ns/op           960 B/op          4 allocs/op
BenchmarkGetCombinations/consumers-10                    1819192               710.9 ns/op          2112 B/op          5 allocs/op
BenchmarkGetCombinations/consumers-10                    1795992               610.2 ns/op          2112 B/op          5 allocs/op
BenchmarkGetCombinations/consumers-10                    1989704               689.6 ns/op          2112 B/op          5 allocs/op
BenchmarkGetCombinations/consumers-10                    1794462               665.6 ns/op          2112 B/op          5 allocs/op
BenchmarkGetCombinations/consumers-10                    1986769               648.5 ns/op          2112 B/op          5 allocs/op
PASS
ok      github.com/kong/kubernetes-ingress-controller/v3/internal/util  17.853s
```

new 9b1af75aadd98aca9550d02518c6dace483a34cb

```
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/util
BenchmarkGetCombinations/consumer_groups-10              7226517               165.5 ns/op           512 B/op          1 allocs/op
BenchmarkGetCombinations/consumer_groups-10              7329960               162.8 ns/op           512 B/op          1 allocs/op
BenchmarkGetCombinations/consumer_groups-10              7136982               161.8 ns/op           512 B/op          1 allocs/op
BenchmarkGetCombinations/consumer_groups-10              7452933               160.4 ns/op           512 B/op          1 allocs/op
BenchmarkGetCombinations/consumer_groups-10              7399339               161.7 ns/op           512 B/op          1 allocs/op
BenchmarkGetCombinations/consumers-10                    4639536               253.1 ns/op           896 B/op          1 allocs/op
BenchmarkGetCombinations/consumers-10                    4779056               256.3 ns/op           896 B/op          1 allocs/op
BenchmarkGetCombinations/consumers-10                    4606189               250.9 ns/op           896 B/op          1 allocs/op
BenchmarkGetCombinations/consumers-10                    4803146               252.7 ns/op           896 B/op          1 allocs/op
BenchmarkGetCombinations/consumers-10                    4661347               253.2 ns/op           896 B/op          1 allocs/op
PASS
ok      github.com/kong/kubernetes-ingress-controller/v3/internal/util  14.439s
```

```
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/util
                                   │   old.txt   │              new.txt               │
                                   │   sec/op    │   sec/op     vs base               │
GetCombinations/consumer_groups-10   342.3n ± 1%   161.8n ± 1%  -52.73% (p=0.008 n=5)
GetCombinations/consumers-10         665.6n ± 4%   253.1n ± 0%  -61.97% (p=0.008 n=5)
geomean                              477.3n        202.4n       -57.60%

                                   │   old.txt    │              new.txt              │
                                   │     B/op     │    B/op     vs base               │
GetCombinations/consumer_groups-10     960.0 ± 0%   512.0 ± 0%  -46.67% (p=0.008 n=5)
GetCombinations/consumers-10          2112.0 ± 0%   896.0 ± 0%  -57.58% (p=0.008 n=5)
geomean                              1.391Ki        677.3       -52.43%

                                   │  old.txt   │              new.txt              │
                                   │ allocs/op  │ allocs/op   vs base               │
GetCombinations/consumer_groups-10   4.000 ± 0%   1.000 ± 0%  -75.00% (p=0.008 n=5)
GetCombinations/consumers-10         5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.008 n=5)
geomean                              4.472        1.000       -77.64%
```

**Which issue this PR fixes**:

<!--
Here you can add any links to issues that this PR is relevant for.
You can use Github keywords (like: closes, fixes or resolves) to auto-resolve
the linked issue(s) when this PR gets merged.

For example: fixes #<issue number>
-->

**Special notes for your reviewer**:

<!-- Here you can add any open questions or notes that you might have for reviewers -->

**PR Readiness Checklist**:

Complete these before marking the PR as `ready to review`:

- [ ] the `CHANGELOG.md` release notes have been updated to reflect any significant (and particularly user-facing) changes introduced by this PR